requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai>=1.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
